refactoring widget components
