--- power-the-polls-form/src/components/address-input/readme.md
@@ -1,28 +1,0 @@
-# address-input
-
-
-
-<!-- Auto Generated Below -->
-
-
-## Dependencies
-
-### Used by
-
- - [power-the-polls-form](../power-the-polls-form)
-
-### Depends on
-
-- [possibly-hidden-select](../possibly-hidden-select)
-
-### Graph
-```mermaid
-graph TD;
-  address-input --> possibly-hidden-select
-  power-the-polls-form --> address-input
-  style address-input fill:#f9f,stroke:#333,stroke-width:4px
-```
-
-----------------------------------------------
-
-
