+++ power-the-polls-form/src/components/ptp-info-jurisdiction/readme.md
@@ -21,12 +21,12 @@
 
 ### Depends on
 
-- [loading-spinner](../loading-spinner)
+- [ui-loading-spinner](../ui-loading-spinner)
 
 ### Graph
 ```mermaid
 graph TD;
-  ptp-info-jurisdiction --> loading-spinner
+  ptp-info-jurisdiction --> ui-loading-spinner
   ptp-info-poll-worker --> ptp-info-jurisdiction
   style ptp-info-jurisdiction fill:#f9f,stroke:#333,stroke-width:4px
 ```
