+++ power-the-polls-form/src/components/ptp-info-poll-worker/readme.md
@@ -32,8 +32,8 @@
 graph TD;
   ptp-info-poll-worker --> ptp-info-jurisdiction
   ptp-info-poll-worker --> ptp-info-state
-  ptp-info-jurisdiction --> loading-spinner
-  ptp-info-state --> loading-spinner
+  ptp-info-jurisdiction --> ui-loading-spinner
+  ptp-info-state --> ui-loading-spinner
   power-the-polls-form --> ptp-info-poll-worker
   style ptp-info-poll-worker fill:#f9f,stroke:#333,stroke-width:4px
 ```
