+++ site/src/components/page-form/readme.md
@@ -23,12 +23,12 @@
 graph TD;
   page-form --> power-the-polls-form
   power-the-polls-form --> ptp-info-poll-worker
-  power-the-polls-form --> address-input
+  power-the-polls-form --> input-address
   ptp-info-poll-worker --> ptp-info-jurisdiction
   ptp-info-poll-worker --> ptp-info-state
-  ptp-info-jurisdiction --> loading-spinner
-  ptp-info-state --> loading-spinner
-  address-input --> possibly-hidden-select
+  ptp-info-jurisdiction --> ui-loading-spinner
+  ptp-info-state --> ui-loading-spinner
+  input-address --> input-possibly-hidden-select
   style page-form fill:#f9f,stroke:#333,stroke-width:4px
 ```
