+++ site/src/components/page-info/readme.md
@@ -19,8 +19,8 @@ graph TD;
   page-info --> ptp-info-poll-worker
   ptp-info-poll-worker --> ptp-info-jurisdiction
   ptp-info-poll-worker --> ptp-info-state
-  ptp-info-jurisdiction --> loading-spinner
-  ptp-info-state --> loading-spinner
+  ptp-info-jurisdiction --> ui-loading-spinner
+  ptp-info-state --> ui-loading-spinner
   style page-info fill:#f9f,stroke:#333,stroke-width:4px
 ```
