+++ site/src/components/page-search/readme.md
@@ -17,14 +17,14 @@
 ### Depends on
 
 - stencil-route-link
-- address-input
+- input-address
 
 ### Graph
 ```mermaid
 graph TD;
   page-search --> stencil-route-link
-  page-search --> address-input
-  address-input --> possibly-hidden-select
+  page-search --> input-address
+  input-address --> input-possibly-hidden-select
   style page-search fill:#f9f,stroke:#333,stroke-width:4px
 ```
